requirements.txt
django>=4.0.0,<4.2.0
djangorestframework
mysqlclient
django-cors-headers
python-dotenv
djangorestframework-simplejwt